still need to fix the sending of emails button
